requirements.txt
# Core Dependencies
#python>=3.8

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning & NLP
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.0.0

# LLM & Agent Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# UI
gradio>=4.0.0
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
loguru>=0.7.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Optional: Document Processing (for Document Manager)
# PyPDF2>=3.0.0  # PDF file support
# python-docx>=0.8.11  # DOCX file support

# Optional: Database Connectors (for Data Pipeline)
# sqlalchemy>=2.0.0  # SQL database ORM
# psycopg2-binary>=2.9.0  # PostgreSQL connector
# pymongo>=4.5.0  # MongoDB connector
# pymysql>=1.1.0  # MySQL connector

# Optional: Feature Store Cache (for Feature Store with Redis)
# redis>=5.0.0  # Redis client for distributed cache

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
ipython>=8.0.0
jupyter>=1.0.0
